Remove `AbstractConfigurableComponent` from `HSMRegistrationClientComponent` and remove `AbstractConfigurableComponent` and `DependenciesTracker`
